REFACTOR: Improve the MVVM file structure
Modified the location of each files and improved the structure to work
better following the MVVM architecture.
