+++ app/src/main/java/com/example/stylish/model/main/BottomNavigationItem.kt
@@ -1,0 +1,10 @@
+package com.example.stylish.model.main
+
+import androidx.compose.ui.graphics.vector.ImageVector
+
+data class BottomNavigationItem(
+    val title: String,
+    val selectedIcon: ImageVector,
+    val unselectedIcon: ImageVector,
+    val route: String
+)
