+++ app/src/main/java/com/example/stylish/presentation/main/MainScreen.kt
@@ -1,0 +1,42 @@
+package com.example.stylish.presentation.main
+
+import androidx.compose.material3.NavigationBar
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.currentBackStackEntryAsState
+import androidx.navigation.compose.rememberNavController
+import com.example.stylish.graphs.MainNavGraph
+import com.example.stylish.presentation.main.component.AddItem
+
+@Composable
+fun MainScreen(
+    navController: NavHostController = rememberNavController(),
+    viewModel: MainViewModel = hiltViewModel()
+) {
+
+    val navBackStackEntry by navController.currentBackStackEntryAsState()
+    val currentDestination = navBackStackEntry?.destination
+
+    Scaffold(
+        bottomBar = {
+            NavigationBar {
+                viewModel.items.forEach { screen ->
+                    AddItem(
+                        screen = screen,
+                        currentDestination = currentDestination,
+                        navController = navController,
+                        navigate = viewModel::navigate
+                    )
+                }
+            }
+        },
+    ) {
+        MainNavGraph(
+            navController = navController,
+            paddingValues = it
+        )
+    }
+}
